refactor: move `strategies` to `helpers/value`
- change named error to be generic as it is not just used by mixins
- remove symbol value mapping from st-var feature: call fields directly
